requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
flask==3.1.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlalchemy==2.0.48
typing-extensions==4.15.0
werkzeug==3.1.6
